requirements.txt
openai
tiktoken
langchain-community
python-dotenv
pinecone-client
langchain-pinecone
clipboard==0.0.4
langchain==0.2.1
langchain_openai==0.1.8
langchain_pinecone==0.1.1
pypdf==4.2.0
python-dotenv==1.0.1
streamlit==1.29.0
unstructured==0.11.5
# sentence-transformers

